fix !, v! and p! - simplify view, color-hl err
this consolidates a number of fixes and adds p! for spawning into a
working pty. this is not entirely enough to run clients like ssh as the
prompt for password breaks on buffering mismatch.

view command now just accepts a number of keywords in any order.

the 'cleanup on finished command with no output' is changed to switch
to err if there is error results back, and a non-0 exit code will set
the job-bar accordingly.