refactor - split out builtins/suggest
This now requires that not only cat9.lua is linked into the lash dir,
but that the subdir is as well. There are more splits to be had now that
the bandaid has been ripped, but the cat9.lua was growing at a troubling
pace. The splits are intended to make it much faster to find where and
what to patch to change some undesired behaviour.

The split of builtins/suggest along with a chainloader (e.g.
cat9/name.lua + cat9/name/command1.lua) means that only the name.lua
needs to be modified to mask out commands, but also that we can have
different interchangeable sets of builtins.

That makes some sense going forward in that switching between different
dedicated cli frontends for complex tools (gdb, ffmpeg, wpa_supplicant,
...) should mainly be a question of swapping out builtin- command sets.

It is still not sufficient for the higher level parsing with specialised
behaviours for tokens, something else is required for that in order to
say - have command-set + token-behaviour emulate classic shells.

FossilOrigin-Name: d2ef8e382bb4f039c467b820580fc2fb117215538814b4f2a5c9c5bd2b793354