add support for views and multicolumn
This was one of the last major complex things on the list. It reworks
the drawing / layouting to let there be a 'per job' view and assigned
column. The view determines higher level presentation, e.g. highlights,
dynamic wrapping or even state machine decoding like vt100.

It also tries to split things into a main 'guaranteed/fill' column, and
when there are more jobs it is subdivided into smaller side columns.