parsing fixes, copy #id(range) working
This adds subrange and per-line picking to some functions, as well
as cleaning up a lot of nuance around parg parsing. It is done by
adding a job:slice(list) method that returns a subselection of the
data contents of the job.

The plan is to use the same setup to get pattern based subsel as
well as data-click actions. Some more polish to the basic commands
along with the vt100 state machine and things are getting to a
decent place.