vt100-csi-sgr working
approaching a tolerable position for the last big boilerplate fest,
being able to attach / detach a terminal state machine to any job.

while supposed to become a configurable option with the wrap view,
the default for view #n wrap is not to at least process sgr attrs
which should give working colors and formatting attributes.

Beneath it all, it also means that the processing and basic vt100
state machine is doing its job though, so adding more csi, osi and
other sequences should be much less effort - especially given that
tui already has 1:1 matching functions for most of it.

this also adds an option of setting a 'default if present' view
mode for imported jobs.