add cd f (favourite) and track dir
Whenever a command is executed from a directory, the directory is
marked as a possible favourite. when that exceeds a certain value,
the directory is returned as part of completion for cd f <dir>.

This is also persisted between runs.

FossilOrigin-Name: aac1c0b293b9c9d3e40255890bd7b06f042066e26788952055b770031e34dd53