add config #job alias
This lets jobs be named and referenced by both name and id. This is
more valuable when a job is also marked persistent, as its id can
change on collision when restoring state. Alias collisions are still
permitted, but any action will match creation order and stop after
first.

It might be useful to permit parg for job creation to also set
alias, something like:

    (alias test, in #0) ls /usr

Though we need to sort the explicit "in as copy" first.