add fallback to builtins
This lets a set of builtins control what happens when there is not
a matching builtin. The default is to setup shell jobs, but for
buitlins that frontend other CLIs it should likely be a forward to
the blobio in charge of the nested CLI process.

FossilOrigin-Name: 3815b0f5672a5ef8b4c6e4ceacdb387c889d0b8d73d2dd82d628066d032cd7d1