refactoring / fixes, line picking
This moves most mouse bindings to the keybindings file, which is
renamed bindings. Switching views is now done through an accessor
that ensures the right state is being tracked/switched. Slicing
data from a view should work properly, and the set of selected
lines is tracked and drawn per view. Mouse 'on-over' indications
for line-numbers have been added, and bindings can now refer to
the currently picked line data index.