support interactive filter defition
this one is great fun, just:

    view #job filter

and it will switch the readline to interactively feed the line into
the filter, letting your figure it out interactively. The next step
for this would be to also have an external filter oracle, such as a
regex engine from another language and asynch feed through that.