add prefix/suffix/group symbols to prompt definition
This should make it slightly less painful to group different
data sources together when defining a prompt, and data sources that
do not yield useful output can be omitted more easily. This is to
prepare for having the prompt be dynamically populated with sources
defined dynamically based on the current set of builtins while
maintaining the same style, e.g [ mysource ] [ another source ].

One example of this would be the state of the currently used wifi
networks, as part of the network builtin with wifi_wpa as the first
provider.