open with media embed follow layout
There are still some visual glitches left, one being the prompt not
refreshing immediately after expand (propagation delay?), another
that the initial size is incorrect and requires a forced relayout
to synch. Otherwise the last missing piece is to respond to the
hidden view state change and hide the job accordingly so that the
WM is free to decompose/detach and even destroy the embedder while
the embedded continues.

That might also require the phandover to allow double-fork, which
should be fixed with privsep setup anyhow.