split builtin/default into default and system
This is a significant structural change. The default builtin set
will now always be loaded first on a set switch, and the ones in
the new set will overlay on that (unless explicitly told not to).

Then a 'system' set of builtins is created for the 'classic' shell
likes of env, cd and so on.

The purpose of this is to shortcircuit the pending problem of WM
and management functions being useful in many builtin sets, but
had to be manually included in each set definition prior to this.

At the same time some system commands have common namespace
collisions that would be messy in other builtin sets, especially
ones that has a _default entry to forward commands to an external
handler.

FossilOrigin-Name: ad2e690d9e8d22dbe2e9fc4724993c79b8c26ddfb11c9a9b83e4473c24f75a93