add helper for reading from external helper
This adds a function that hooks into the timer subsystem to implement
a reader from a continously streaming source. The callback is
triggered when there hasn't been any new data on the source for a
certain amount of ticks (ticks * 25 ms) or if the source stream
dies while reading, then reschedules itself.

cat9.reader_factory(stream, ticks_timeout,
	function(blob, dead)
	end
)

The purpose is to deal with wrapping CLI tools that keep running
and are clocked by some external source without blocking / making
cat9 unresponsive.